refactor: move test_utils module
Previously this module was located specifically under the `section_authority_provider` module, but
now it moves to its own module and various DBC-related testing utilities are moved from `sn_node`
to this location. Again, this test setup is needed by both `sn_node` and `sn_client`.
