+++ sn_interface/src/network_knowledge/section_authority_provider.rs
@@ -271,179 +271,3 @@ impl SectionAuth<SectionAuthorityProviderMsg> {
         }
     }
 }
-
-#[cfg(any(test, feature = "test-utils"))]
-pub mod test_utils {
-    use super::*;
-    use crate::messaging::system::{KeyedSig, SectionAuth};
-    use crate::network_knowledge::{Error, NodeInfo, MIN_ADULT_AGE};
-    use eyre::{eyre, Result};
-    use itertools::Itertools;
-    use rand::RngCore;
-    use serde::Serialize;
-    use sn_consensus::{Ballot, Consensus, Decision, Proposition, Vote, VoteResponse};
-    use std::collections::BTreeMap;
-    use std::{cell::Cell, net::SocketAddr};
-    use xor_name::Prefix;
-
-    // Parse `Prefix` from string
-    pub fn prefix(s: &str) -> Result<Prefix> {
-        s.parse()
-            .map_err(|err| eyre!("failed to parse Prefix '{}': {}", s, err))
-    }
-
-    // Generate unique SocketAddr for testing purposes
-    pub fn gen_addr() -> SocketAddr {
-        thread_local! {
-            static NEXT_PORT: Cell<u16> = Cell::new(1000);
-        }
-
-        let port = NEXT_PORT.with(|cell| cell.replace(cell.get().wrapping_add(1)));
-
-        ([192, 0, 2, 0], port).into()
-    }
-
-    // Create `count` Nodes sorted by their names.
-    // The `age_diff` flag is used to trigger nodes being generated with different age pattern.
-    // The test of `handle_agreement_on_online_of_elder_candidate` requires most nodes to be with
-    // age of MIN_AGE + 2 and one node with age of MIN_ADULT_AGE.
-    pub fn gen_sorted_nodes(prefix: &Prefix, count: usize, age_diff: bool) -> Vec<NodeInfo> {
-        (0..count)
-            .map(|index| {
-                let age = if age_diff && index < count - 1 {
-                    MIN_ADULT_AGE + 1
-                } else {
-                    MIN_ADULT_AGE
-                };
-                NodeInfo::new(
-                    crate::types::keys::ed25519::gen_keypair(&prefix.range_inclusive(), age),
-                    gen_addr(),
-                )
-            })
-            .sorted_by_key(|node| node.name())
-            .collect()
-    }
-
-    /// Generate a random `SectionAuthorityProvider` for testing.
-    ///
-    /// The total number of members in the section will be `elder_count` + `adult_count`. A lot of
-    /// tests don't require adults in the section, so zero is an acceptable value for
-    /// `adult_count`.
-    ///
-    /// An optional `sk_threshold_size` can be passed to specify the threshold when the secret key
-    /// set is generated for the section key. Some tests require a low threshold.
-    pub fn random_sap_with_rng<R: RngCore>(
-        rng: &mut R,
-        prefix: Prefix,
-        elder_count: usize,
-        adult_count: usize,
-        sk_threshold_size: Option<usize>,
-    ) -> (SectionAuthorityProvider, Vec<NodeInfo>, bls::SecretKeySet) {
-        let nodes = gen_sorted_nodes(&prefix, elder_count + adult_count, false);
-        let elders = nodes.iter().map(NodeInfo::peer).take(elder_count);
-        let members = nodes.iter().map(|i| NodeState::joined(i.peer(), None));
-        let poly = bls::poly::Poly::random(sk_threshold_size.unwrap_or(0), rng);
-        let sks = bls::SecretKeySet::from(poly);
-        let section_auth =
-            SectionAuthorityProvider::new(elders, prefix, members, sks.public_keys(), 0);
-        (section_auth, nodes, sks)
-    }
-
-    pub fn random_sap(
-        prefix: Prefix,
-        elder_count: usize,
-        adult_count: usize,
-        sk_threshold_size: Option<usize>,
-    ) -> (SectionAuthorityProvider, Vec<NodeInfo>, bls::SecretKeySet) {
-        random_sap_with_rng(
-            &mut rand::thread_rng(),
-            prefix,
-            elder_count,
-            adult_count,
-            sk_threshold_size,
-        )
-    }
-
-    /// Generate a random `SectionAuthorityProvider` for testing.
-    ///
-    /// The same as `random_sap`, but instead the secret key is provided. This can be useful for
-    /// creating a section to share the same genesis key as another one.
-    pub fn random_sap_with_key(
-        prefix: Prefix,
-        elder_count: usize,
-        adult_count: usize,
-        sk_set: &bls::SecretKeySet,
-    ) -> (SectionAuthorityProvider, Vec<NodeInfo>) {
-        let nodes = gen_sorted_nodes(&prefix, elder_count + adult_count, false);
-        let elders = nodes.iter().map(NodeInfo::peer).take(elder_count);
-        let members = nodes.iter().map(|i| NodeState::joined(i.peer(), None));
-        let section_auth =
-            SectionAuthorityProvider::new(elders, prefix, members, sk_set.public_keys(), 0);
-        (section_auth, nodes)
-    }
-
-    // Create signature for the given payload using the given secret key.
-    pub fn prove<T: Serialize>(secret_key: &bls::SecretKey, payload: &T) -> Result<KeyedSig> {
-        let bytes = bincode::serialize(payload).map_err(|_| Error::InvalidPayload)?;
-        Ok(KeyedSig {
-            public_key: secret_key.public_key(),
-            signature: secret_key.sign(&bytes),
-        })
-    }
-
-    pub fn section_decision<P: Proposition>(
-        secret_key_set: &bls::SecretKeySet,
-        proposal: P,
-    ) -> Result<Decision<P>> {
-        let n = secret_key_set.threshold() + 1;
-        let mut nodes = Vec::from_iter((1..=n).into_iter().map(|idx| {
-            let secret = (idx as u8, secret_key_set.secret_key_share(idx));
-            Consensus::from(secret, secret_key_set.public_keys(), n)
-        }));
-
-        let first_vote = nodes[0].sign_vote(Vote {
-            gen: 0,
-            ballot: Ballot::Propose(proposal),
-            faults: Default::default(),
-        })?;
-
-        let mut votes = vec![nodes[0].cast_vote(first_vote)?];
-
-        while let Some(vote) = votes.pop() {
-            for node in &mut nodes {
-                match node.handle_signed_vote(vote.clone())? {
-                    VoteResponse::WaitingForMoreVotes => (),
-                    VoteResponse::Broadcast(vote) => votes.push(vote),
-                }
-            }
-        }
-
-        // All nodes have agreed to the same proposal
-        assert_eq!(
-            BTreeSet::from_iter(nodes.iter().map(|n| {
-                if let Some(d) = n.decision.clone() {
-                    d.proposals
-                } else {
-                    BTreeMap::new()
-                }
-            }))
-            .len(),
-            1
-        );
-
-        let decision = nodes[0]
-            .decision
-            .clone()
-            .ok_or_else(|| eyre!("A decision was expected to be found for this particular node"))?;
-        Ok(decision)
-    }
-
-    // Wrap the given payload in `SectionAuth`
-    pub fn section_signed<T: Serialize>(
-        secret_key: &bls::SecretKey,
-        payload: T,
-    ) -> Result<SectionAuth<T>> {
-        let sig = prove(secret_key, &payload)?;
-        Ok(SectionAuth::new(payload, sig))
-    }
-}
